Pull process_html() into a module so it can be reused elsewhere
